pyproject.toml
[tool.poetry]
name = "spotify_to_plex"
version = "2.0.0"
description = "Sync Spotify playlists to Plex Media Server"
authors = ["lammersbjorn <bjorn@lammers.media>", "0xChloe <chloe@catastrophesoftware.com>"]
license = "GPLv3"
readme = "README.md"
repository = "https://github.com/lammersbjorn/spotify-to-plex"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
PlexAPI = "^4.15.13"
spotipy = "^2.24.0"
rtoml = "^0.10.0"
typer = "^0.12.3"
rich = "^13.7.1"
loguru = "^0.7.2"
python-dotenv = "^1.1.0"
tqdm = "^4.67.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
ruff = "^0.1.0"
mypy = "^1.6.0"
pytest = "^7.4.2"

[tool.poetry.scripts]
spotify-to-plex = "spotify_to_plex.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["ALL"]
ignore = ["D203", "D212"]
target-version = "py310"
line-length = 100

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "D104"]
"tests/*" = ["D100", "D101", "D102", "D103"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
